Extract terminal window appearance into window_appearance.rs (E1 tranche 1).
Move background blur/opacity resolution, GPUI window background sync, and system
appearance handling out of mod.rs with tests co-located in the new module.
